+++ src/pages/campaign/landing/components/table/table-column-amount.tsx
