+++ src/pages/campaign/step/index.tsx
@@ -1,77 +1,22 @@
-import { useMemo } from 'react';
 import tw from 'twin.macro';
 
-import { IconBack, IconNext } from '~/assets/icons';
-
-import { ButtonPrimarySmallIconLeading } from '~/components/buttons/primary/small-icon-leading';
-import { ButtonPrimarySmallIconTrailing } from '~/components/buttons/primary/small-icon-trailing';
-
-import { useConnectedWallet } from '~/hooks/wallets';
-
-import { LoadingStep } from '../components/stepper';
 import { StepContents } from '../layouts/layout-step';
-import { useCampaignStepStore } from '../states/step';
+
+import { Stepper } from './components/stepper';
+import { StepperTitle } from './components/title';
 
 const StepPage = () => {
-  const { step, setStep, isLoading } = useCampaignStepStore();
-  const { xrp, evm } = useConnectedWallet();
-
-  const network = step === 1 ? 'XRPL' : 'the Root Network';
-  const titleText =
-    step === 1 || step === 2
-      ? isLoading
-        ? 'Waiting to connect...'
-        : `Connect to ${network} wallet`
-      : step === 3
-      ? 'Bridge your $XRP'
-      : 'Add liquidity';
-
-  const nextDisabled = useMemo(() => {
-    if (step === 1) {
-      return !xrp.isConnected;
-    }
-    if (step === 2) {
-      return !evm.isConnected;
-    }
-    // TODP : add condition step3, step4
-  }, [evm.isConnected, step, xrp.isConnected]);
-
   return (
     <Wrapper>
-      <LoadingStep />
-      <Content>
-        {(step !== 1 || xrp.isConnected) && (
-          <MoveStepWrapper>
-            <ButtonPrimarySmallIconLeading
-              text={'Back'}
-              icon={<IconBack />}
-              disabled={step === 1}
-              onClick={() => setStep('negative')}
-            />
-            <ButtonPrimarySmallIconTrailing
-              text={'Next'}
-              icon={<IconNext />}
-              disabled={nextDisabled || step === 4}
-              onClick={() => setStep('positive')}
-            />
-          </MoveStepWrapper>
-        )}
+      <Stepper />
 
-        <TextWrapper>
-          <StepText>Step {step}</StepText>
-          <Title>{titleText}</Title>
-        </TextWrapper>
+      <StepperTitle />
 
-        <StepContents />
-      </Content>
+      <StepContents />
     </Wrapper>
   );
 };
 
-const Wrapper = tw.div`flex flex-col items-center p-80 gap-40`;
-const Content = tw.div`w-455 flex flex-col items-center gap-24`;
-const MoveStepWrapper = tw.div`w-full flex justify-between items-center`;
-const TextWrapper = tw.div`w-full flex flex-col gap-4`;
-const StepText = tw.div`font-m-14 text-primary-60`;
-const Title = tw.div`font-b-24 text-neutral-100`;
+const Wrapper = tw.div`w-455 flex-center flex-col py-80 gap-24`;
+
 export default StepPage;
